# Angkutmi-Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Angkutmi API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/payments": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Membuat pembayaran dan proses pembayaran",
                "description": "Menginput data pembayaran baru dan menyimpan data proses pembayaran",
                "operationId": "5e2794ba3d3c64e70873367505e4a751",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "trip_id",
                                    "payment_type",
                                    "base_price",
                                    "final_price"
                                ],
                                "properties": {
                                    "trip_id": {
                                        "description": "ID Trip yang telah selesai",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "payment_type": {
                                        "description": "Tipe pembayaran",
                                        "type": "string",
                                        "enum": [
                                            "debit",
                                            "ovo"
                                        ],
                                        "example": "debit"
                                    },
                                    "base_price": {
                                        "description": "Harga awal",
                                        "type": "integer",
                                        "example": 50000
                                    },
                                    "final_price": {
                                        "description": "Harga akhir setelah diskon",
                                        "type": "integer",
                                        "example": 45000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pembayaran berhasil diproses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pembayaran berhasil diproses."
                                        },
                                        "payment": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "payment_type": {
                                                    "type": "string",
                                                    "example": "debit"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-06-15T10:00:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-06-15T10:00:00"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "payment_process": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "trip_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "payment_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "base_price": {
                                                    "type": "integer",
                                                    "example": 50000
                                                },
                                                "final_price": {
                                                    "type": "integer",
                                                    "example": 45000
                                                },
                                                "payment_status": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-06-15T10:00:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-06-15T10:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/payments/{trip_id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Melihat detail pembayaran berdasarkan trip ID",
                "description": "Mengambil data proses pembayaran dan pembayaran berdasarkan trip ID",
                "operationId": "0004d3261fc59bb4a3f151514a0fad5b",
                "parameters": [
                    {
                        "name": "trip_id",
                        "in": "path",
                        "description": "ID Trip yang ingin dilihat detail pembayarannya",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pembayaran berhasil ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "payment_process": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "trip_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "payment_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "base_price": {
                                                    "type": "integer",
                                                    "example": 50000
                                                },
                                                "final_price": {
                                                    "type": "integer",
                                                    "example": 45000
                                                },
                                                "payment_status": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "payment": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "payment_type": {
                                                    "type": "string",
                                                    "example": "debit"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "trip": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "is_started": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data pembayaran tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Data pembayaran tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}